"code tu day nha ae"

+++ app/Http/Controllers/TransactionPointHeadController.php
@@ -3,9 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use App\Models\TransactionPointHead;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class TransactionPointHeadController extends Controller
 {
@@ -22,21 +24,39 @@ class TransactionPointHeadController extends Controller
      */
     public function store(Request $request)
     {
-        //
-        $user = User::create([
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => Hash::make($request->password),
-        ]);
-        $user->transactionPointHead()->create([
-            'transaction_point_id' => $request->transaction_point_id,
-            'phone' => $request->phone,
-            'address' => $request->address,
-            'details' => $request->details,
-        ]);
-        return response()->json([
-            'message' => "tao tai khoan thanh cong !",
-        ]);
+        if (Auth::user()->isCompanyLeader()) {
+
+            $validator = Validator::make($request->all(), [
+                'transaction_point_id' => 'required|numeric|unique_transaction_point',
+            ]);
+            if ($validator->fails()) {
+                return response()->json(['errors' => $validator->errors()], 422);
+            } else {
+                $user = User::create([
+                    'name' => $request->name,
+                    'email' => $request->email,
+                    'password' => Hash::make($request->password),
+                ]);
+                $transactionPointHead = $user->transactionPointHead()->create([
+                    'transaction_point_id' => $request->transaction_point_id,
+                    'phone' => $request->phone,
+                    'details' => $request->details,
+                ]);
+                $transactionPointHead->address()->create([
+                    'province' => $request->input('province', 'default $transactionPointHead'),
+                    'district' => $request->input('district', 'default $transactionPointHead'),
+                    'ward' => $request->input('ward', 'default $transactionPointHead'),
+                    'detailed_address' => $request->input('detailed_address', 'default $transactionPointHead'),
+                ]);
+                return response()->json([
+                    'message' => "tao tai khoan thanh cong !",
+                ]);
+            }
+        } else {
+            return response()->json([
+                'message' => "tao tai khoan thanh cong !",
+            ]);
+        }
     }
 
     /**
@@ -66,10 +86,14 @@ class TransactionPointHeadController extends Controller
         $transactionPointHead->update([
             'transaction_point_id' => $request->transaction_point_id,
             'phone' => $request->phone,
-            'address' => $request->address,
             'details' => $request->details,
         ]);
-
+        $transactionPointHead()->address()->create([
+            'province' => $request->input('province', 'default $transactionPointHead'),
+            'district' => $request->input('district', 'default $transactionPointHead'),
+            'ward' => $request->input('ward', 'default $transactionPointHead'),
+            'detailed_address' => $request->input('detailed_address', 'default $transactionPointHead'),
+        ]);
         return response()->json([
             'message' => 'Cập nhật thành công!',
         ]);
@@ -80,15 +104,21 @@ class TransactionPointHeadController extends Controller
      */
     public function destroy(string $id)
     {
-        //
-        $transactionPointHead = TransactionPointHead::with('user')->findOrFail($id);
-        $transactionPointHead->delete();
+        try {
+            $transactionPointHead = TransactionPointHead::with('user')->findOrFail($id);
 
-        // Xóa thông tin trong bảng User
-        $transactionPointHead->user->delete();
-
-        return response()->json([
-            'message' => 'Xóa tài khoản thành công!',
-        ]);
+            if (Auth::user()->isCompanyLeader()) {
+                // Delete the TransactionPointHead record
+                $transactionPointHead->delete();
+                // Delete the associated user record
+                $transactionPointHead->user->delete();
+                // You may also want to delete associated records or perform other cleanup tasks.
+                return response()->json(['message' => 'Xóa tài khoản thành công!']);
+            } else {
+                return response()->json(['error' => 'Không đủ quyền hạn'], 403);
+            }
+        } catch (\Exception $e) {
+            return response()->json(['error' => 'Không thể xóa tài khoản.'], 500);
+        }
     }
 }
